\encoding{utf8} 
\name{bayesZIB-internal}
\alias{bayesZIB-internal}
\alias{Rcpp_rstantools_model_model-class}
\alias{_rcpp_module_boot_stan_fit4model_mod}
\alias{rstantools_model_model}
\alias{stanmodels}

\title{Internal bayesZIB functions and classes}
\description{Internal bayesZIB functions and classes}
\usage{
rstantools_model_model()
stanmodels()
}

\details{These functions are not to be called by the user}
\seealso{
\code{\link{bayesZIB-package}}, \code{\link{bayesZIB}}
}
\keyword{internal}
